Definitions are now held in a separate bundle.
They can now be read by the IDE without the need to include the
definitions file in the project.  Simplifies deployment. This did mean
that all tests now have to work with the production language spec, which
also means that some core language features may not be as well tested as
I'd like. Sadly using both test and prodn definitions did not work in
the mvn build.